requirements.txt
python-dotenv
pandas
instabot
pillow
pytest
